Change connected store to the Table component instead of App component
