+++ src/app.js
@@ -9,13 +9,13 @@ import configureStore from './store/configStore';
 import css from './style.scss';
 
 // import components
-import App from './Components/App';
+import Main from './Components/Main';
 
 const store = configureStore();
 
 render(
   <Provider store={store}>
-    <App />
+    <Main />
   </Provider>,
   document.getElementById(`root`)
 );
